requirements.txt
# Mac OS Zero Trust Endpoint Security Platform
# Author: Adrian Johnson <adrian207@gmail.com>

# Core Framework
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0

# API and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
requests==2.31.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Security and Authentication
cryptography==41.0.7
pyjwt==2.8.0
bcrypt==4.1.1
python-multipart==0.0.6

# Task Queue and Background Jobs
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==23.2.0

# Data Processing and Analysis
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Cloud SDKs (Optional - uncomment as needed)
# boto3==1.29.7  # AWS
# azure-identity==1.15.0  # Azure
# azure-mgmt-compute==30.4.0
# google-cloud-storage==2.10.0  # GCP

# Security Tool SDKs (Mock/Example - replace with actual SDKs)
# kandji-sdk==1.0.0  # Replace with actual package
# zscaler-sdk==1.0.0  # Replace with actual package
# seraphic-sdk==1.0.0  # Replace with actual package

# Mac OS System Integration
pyobjc-core==10.0; sys_platform == 'darwin'
pyobjc-framework-SystemConfiguration==10.0; sys_platform == 'darwin'
pyobjc-framework-SecurityFoundation==10.0; sys_platform == 'darwin'

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3
isort==5.12.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# Utilities
click==8.1.7
rich==13.7.0
tabulate==0.9.0
schedule==1.2.0

